Support for python on Windows and cleanup initialisation.
Remove unnecessary require statements for lsp (now lps-mode)
and markdown - moved to use-package and init.el.

Add support for conda on Windows with pyvenv.
pyvenv doesn't add all the required directories for
conda/miniconda. Redefine the pyvenv-activate function if we're using
Windows and Conda and add all directories required to the path in
order to correctly load binary dependent files.