Support flycheck checker chain with python and latest lsp-mode.
lsp-mode changes have forced a change to my configuration which
supports multiple flycheck checkers with python and mspyls.
When enabling lsp flycheck checker for the mode we check if the mode
is python mode and configure the chain, then set the first checker to
be flake8. The final checker will be the usual default lsp.