Update run current buffer and pytrder stats.
When running commands check if a buffer with the same name is being
used and close the buffer, then run the new command.

pytrader-stats (and debug version) now kill the server if it's running
and restart. This allows quick development, allowing me to add new
code and then bounce and restart the server quickly.